refactor(readpath): extract degraded-mode broadcast into a Broadcaster
readpath.Failover handled both the normal DB-backed failover loop and the
entire degraded-mode broadcast machinery. Split the broadcast behavior into
a dedicated Broadcaster (new broadcaster.go): it owns the cache check,
sequential/parallel fan-out, loser cancel + drain, and the winner
cache-write. Failover keeps the policy layer - resolve locations, try each
copy, and the degraded gate (degraded-mode span attr, DegradedReadsTotal,
the degradedReadsEnabled check) - then delegates to broadcaster.Read on a
DB outage. No behavior change; existing read-path/integration tests pass.

Also corrected two stale references in the read-path diagram: the failover
node named a removed withReadFailover() function, and the broadcast node
described loser cleanup as sync.Once when it is context-cancel + drain.

Closes #977
